Extract cleanAST to language-specific files
