+++ src/language-css/printer-postcss.js
@@ -1,5 +1,6 @@
 "use strict";
 
+const clean = require("./clean");
 const util = require("../common/util");
 const doc = require("../doc");
 const docBuilders = doc.builders;
@@ -535,5 +536,6 @@ function maybeToLowerCase(value) {
 
 module.exports = {
   print: genericPrint,
-  hasPrettierIgnore: util.hasIgnoreComment
+  hasPrettierIgnore: util.hasIgnoreComment,
+  massageAstNode: clean
 };
