+++ src/language-markdown/printer-markdown.js
@@ -658,8 +658,20 @@ function normalizeParts(parts) {
   }, []);
 }
 
+function clean(ast, newObj) {
+  // for markdown codeblock
+  if (ast.type === "code") {
+    delete newObj.value;
+  }
+  // for markdown whitespace: "\n" and " " are considered the same
+  if (ast.type === "whitespace" && ast.value === "\n") {
+    newObj.value = " ";
+  }
+}
+
 module.exports = {
   print: genericPrint,
   embed,
+  massageAstNode: clean,
   hasPrettierIgnore: util.hasIgnoreComment
 };
