+++ src/language-vue/index.js
@@ -21,12 +21,18 @@ const languages = [
   }
 ];
 
+const clean = (ast, newObj) => {
+  delete newObj.contentStart;
+  delete newObj.contentEnd;
+};
+
 const parsers = {
   vue: {
     get parse() {
       return eval("require")("./parser-vue");
     },
-    astFormat: "vue"
+    astFormat: "vue",
+    massageAstNode: clean
   }
 };
